Revert Bar.tsx back to Bar.js
